+++ noodles-sam/src/reader/record/sequence.rs
@@ -8,6 +8,8 @@ use crate::record::{
 /// An error when a raw SAM record sequence fails to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ParseError {
+    /// The input is empty.
+    Empty,
     /// A base is invalid.
     InvalidBase(base::TryFromCharError),
 }
@@ -15,6 +17,7 @@ pub enum ParseError {
 impl error::Error for ParseError {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         match self {
+            Self::Empty => None,
             Self::InvalidBase(e) => Some(e),
         }
     }
@@ -23,12 +26,17 @@ impl error::Error for ParseError {
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
+            Self::Empty => write!(f, "empty input"),
             Self::InvalidBase(_) => write!(f, "invalid base"),
         }
     }
 }
 
 pub(crate) fn parse_sequence(src: &[u8], sequence: &mut Sequence) -> Result<(), ParseError> {
+    if src.is_empty() {
+        return Err(ParseError::Empty);
+    }
+
     let mut bases = Vec::from(mem::take(sequence));
 
     for &n in src {
@@ -50,14 +58,12 @@ mod tests {
         let mut sequence = Sequence::default();
 
         sequence.clear();
-        parse_sequence(b"", &mut sequence)?;
-        let expected = Sequence::default();
-        assert_eq!(sequence, expected);
-
-        sequence.clear();
         parse_sequence(b"ACGT", &mut sequence)?;
         let expected = Sequence::from(vec![Base::A, Base::C, Base::G, Base::T]);
         assert_eq!(sequence, expected);
+
+        sequence.clear();
+        assert_eq!(parse_sequence(b"", &mut sequence), Err(ParseError::Empty));
 
         sequence.clear();
         assert!(matches!(